feat: create composable for keyboard navigation
