+++ aform/stories/default.story.vue
@@ -58,7 +58,7 @@ const formSetup = ({ app }: { app: App }) => {
 
 <style>
 @import url('https://fonts.googleapis.com/css2?family=Arimo:ital,wght@0,400;0,500;0,600;0,700;1,400;1,500;1,600;1,700&display=swap');
-@import url('../src/theme/aform.css');
+@import url('@/theme/aform.css');
 
 html {
 	font-family: Arimo, sans-serif;
