+++ aform/tsconfig.json
@@ -16,6 +16,7 @@
 		"src/**/*",
 		"src/**/*.ts",
 		"src/**/*.vue",
+		"src/**/*.json",
 		"stories/**/*.story.vue",
 		"tests/**/*.ts",
 		"tests/**/*.vue",
