+++ atable/src/components/ATable.vue
@@ -47,7 +47,7 @@
 
 <script lang="ts">
 import { v4 } from 'uuid'
-import { defineComponent, nextTick, PropType, provide } from 'vue'
+import { defineComponent, PropType, provide } from 'vue'
 
 import { TableColumn, TableConfig, TableRow } from 'types'
 import TableDataStore from '.'
@@ -131,187 +131,6 @@ export default defineComponent({
 			}
 		}
 
-		const getIndent = (colKey: number, indent: number) => {
-			if (indent && colKey === 0 && indent > 0) {
-				return `${indent}ch`
-			} else {
-				return null
-			}
-		}
-
-		const enterNav = async (event: KeyboardEvent) => {
-			event.preventDefault()
-			event.stopPropagation()
-			event.shiftKey ? await upCell(event) : await downCell(event)
-		}
-
-		const tabNav = async (event: KeyboardEvent) => {
-			event.preventDefault()
-			event.stopPropagation()
-			event.shiftKey ? await prevCell(event) : await nextCell(event)
-		}
-
-		const downArrowNav = async (event: KeyboardEvent) => {
-			if (!event.shiftKey) {
-				event.preventDefault()
-				event.stopPropagation()
-				await downCell(event)
-			}
-		}
-
-		const upArrowNav = async (event: KeyboardEvent) => {
-			if (!event.shiftKey) {
-				event.preventDefault()
-				event.stopPropagation()
-				await upCell(event)
-			}
-		}
-
-		const leftArrowNav = async (event: KeyboardEvent) => {
-			if (!event.shiftKey) {
-				event.preventDefault()
-				event.stopPropagation()
-				await prevCell(event)
-			}
-		}
-
-		const rightArrowNav = async (event: KeyboardEvent) => {
-			if (!event.shiftKey) {
-				event.preventDefault()
-				event.stopPropagation()
-				await nextCell(event)
-			}
-		}
-
-		const endNav = (event: KeyboardEvent) => {
-			const $cell = event.target as HTMLTableCellElement
-			const cellIndex = $cell.cellIndex
-			const $row = $cell.parentElement as HTMLTableRowElement
-			const rowIndex = $row.rowIndex
-			const $table = $row.parentElement as HTMLTableElement
-
-			const $lastRow = $table.rows[rowIndex - 1]
-			if ($lastRow.cells.length - 1 !== cellIndex) {
-				// last column
-				const $nextCell = $lastRow.cells[tableData.columns.length - (tableData.zeroColumn ? 0 : 1)] // last cell
-				$nextCell.focus()
-			}
-		}
-
-		const homeNav = (event: KeyboardEvent) => {
-			const $cell = event.target as HTMLTableCellElement
-			const cellIndex = $cell.cellIndex
-			const $row = $cell.parentElement as HTMLTableRowElement
-			const rowIndex = $row.rowIndex
-			const $table = $row.parentElement as HTMLTableElement
-
-			const $lastRow = $table.rows[rowIndex - 1]
-			if (cellIndex !== (tableData.config.numberedRows ? 1 : 0)) {
-				// TODO: zeroColumn // first column
-				const $nextCell = $lastRow.cells[tableData.zeroColumn ? 1 : 0] // last cell
-				$nextCell.focus()
-			}
-		}
-
-		const downCell = async (event: KeyboardEvent) => {
-			const $cell = event.target as HTMLTableCellElement
-			const cellIndex = $cell.cellIndex
-			const $row = $cell.parentElement as HTMLTableRowElement
-			const rowIndex = $row.rowIndex
-			const $table = $row.parentElement as HTMLTableElement
-
-			let $nextCell = event.target as HTMLTableCellElement
-			if ($table.rows.length !== rowIndex) {
-				// not bottom row
-				$nextCell = $table.rows[rowIndex].cells[cellIndex]
-				if (tableData.config.treeView && !tableData.display[rowIndex].open) {
-					// toggleRowExpand(event, rowIndex - 1)
-					tableData.toggleRowExpand(rowIndex - 1)
-				}
-			}
-
-			await nextTick()
-			$nextCell.focus()
-		}
-
-		const upCell = async (event: KeyboardEvent) => {
-			const $cell = event.target as HTMLTableCellElement
-			const cellIndex = $cell.cellIndex
-			const $row = $cell.parentElement as HTMLTableRowElement
-			const rowIndex = $row.rowIndex
-			const $table = $row.parentElement as HTMLTableElement
-
-			let $nextCell = event.target as HTMLTableCellElement
-			if (rowIndex !== 1) {
-				// not top row, exclude headers
-				$nextCell = $table.rows[rowIndex - 2].cells[cellIndex]
-				if (tableData.config.treeView && !tableData.display[rowIndex - 2].open) {
-					tableData.toggleRowExpand(tableData.display[rowIndex - 2].parent)
-				}
-			}
-
-			await nextTick()
-			$nextCell.focus()
-		}
-
-		const nextCell = async (event: KeyboardEvent) => {
-			const $cell = event.target as HTMLTableCellElement
-			const cellIndex = $cell.cellIndex
-			const $row = $cell.parentElement as HTMLTableRowElement
-			const rowIndex = $row.rowIndex
-			const $table = $row.parentElement as HTMLTableElement
-
-			let $nextCell: HTMLTableCellElement
-			const $lastRow = $table.rows[rowIndex - 1]
-			if ($lastRow.cells.length - 1 === cellIndex) {
-				// last column
-				if ($table.rows.length === rowIndex) {
-					$nextCell = $table.rows[0].cells[tableData.zeroColumn ? 1 : 0] // go to top left cell
-					// if row is hidden, expand
-				} else {
-					// focus on first cell of next row
-					$nextCell = $table.rows[rowIndex].cells[tableData.zeroColumn ? 1 : 0]
-					if (tableData.config.treeView && !tableData.display[rowIndex].open) {
-						tableData.toggleRowExpand(rowIndex - 1)
-					}
-				}
-			} else {
-				$nextCell = $lastRow.cells[cellIndex + 1] // next cell
-			}
-
-			await nextTick()
-			$nextCell.focus()
-		}
-
-		const prevCell = async (event: KeyboardEvent) => {
-			const $cell = event.target as HTMLTableCellElement
-			const cellIndex = $cell.cellIndex
-			const $row = $cell.parentElement as HTMLTableRowElement
-			const rowIndex = $row.rowIndex
-			const $table = $row.parentElement as HTMLTableElement
-
-			let $prevCell: HTMLTableCellElement
-			const $lastRow = $table.rows[rowIndex - 1]
-			const $secondLastRow = $table.rows[rowIndex - 2]
-			if (cellIndex === (tableData.zeroColumn ? 1 : 0)) {
-				// first column
-				if (rowIndex !== 1) {
-					// not top row, exclude headers
-					$prevCell = $secondLastRow.cells[$secondLastRow.cells.length - 1]
-					// toggleRowExpand(event, rowIndex - 2)
-					tableData.toggleRowExpand(rowIndex - 2)
-				} else {
-					// top row, stay trapped in top left cell
-					return
-				}
-			} else {
-				$prevCell = $lastRow.cells[cellIndex - 1] // previous cell
-			}
-
-			await nextTick()
-			$prevCell.focus()
-		}
-
 		const moveCursorToEnd = (target: HTMLElement) => {
 			target.focus()
 			document.execCommand('selectAll', false, null)
@@ -330,22 +149,9 @@ export default defineComponent({
 		window.addEventListener('click', clickOutside)
 
 		return {
-			downArrowNav,
-			downCell,
-			endNav,
-			enterNav,
 			formatCell,
-			getIndent,
-			homeNav,
-			leftArrowNav,
 			moveCursorToEnd,
-			nextCell,
-			prevCell,
-			rightArrowNav,
 			tableData,
-			tabNav,
-			upArrowNav,
-			upCell,
 			v4,
 		}
 	},
