+++ atable/src/index.ts
@@ -6,6 +6,7 @@ import ATable from '@/components/ATable.vue'
 import ATableHeader from '@/components/ATableHeader.vue'
 import ATableModal from '@/components/ATableModal.vue'
 import TableDataStore from './components'
+import { useKeyboardNav } from './composables/keyboard'
 
 function install(app: App /* options */) {
 	app.component('ACell', ACell)
@@ -15,4 +16,4 @@ function install(app: App /* options */) {
 	app.component('ATableModal', ATableModal)
 }
 
-export { install, ACell, ARow, ATable, ATableHeader, ATableModal, TableDataStore }
+export { install, ACell, ARow, ATable, ATableHeader, ATableModal, TableDataStore, useKeyboardNav }
